+++ test/orm/active_record.rb
@@ -7,22 +7,11 @@ ActiveRecord::Base.include_root_in_json = true
 migrate_path = File.expand_path("../../rails_app/db/migrate/", __FILE__)
 if Devise::Test.rails71_and_up?
   ActiveRecord::MigrationContext.new(migrate_path).migrate
-elsif Devise::Test.rails6_and_up?
-  ActiveRecord::MigrationContext.new(migrate_path, ActiveRecord::SchemaMigration).migrate
-elsif Devise::Test.rails52_and_up?
-  ActiveRecord::MigrationContext.new(migrate_path).migrate
 else
-  ActiveRecord::Migrator.migrate(migrate_path)
+  ActiveRecord::MigrationContext.new(migrate_path, ActiveRecord::SchemaMigration).migrate
 end
 
 class ActiveSupport::TestCase
-  if Devise::Test.rails5_and_up?
-    self.use_transactional_tests = true
-  else
-    # Let `after_commit` work with transactional fixtures, however this is not needed for Rails 5.
-    require 'test_after_commit'
-    self.use_transactional_fixtures = true
-  end
-
+  self.use_transactional_tests = true
   self.use_instantiated_fixtures  = false
 end