+++ test/rails_app/config/application.rb
@@ -33,21 +33,10 @@ module RailsApp
     # config.assets.enabled = false
 
     config.action_mailer.default_url_options = { host: "localhost", port: 3000 }
-    rails_version = Gem::Version.new(Rails.version)
-    if DEVISE_ORM == :active_record &&
-       rails_version >= Gem::Version.new('4.2.0') &&
-       rails_version < Gem::Version.new('5.1.0')
-      config.active_record.raise_in_transactional_callbacks = true
-    end
 
     # This was used to break devise in some situations
     config.to_prepare do
       Devise::SessionsController.layout "application"
-    end
-
-    # Remove the first check once Rails 5.0 support is removed.
-    if Devise::Test.rails52_and_up? && !Devise::Test.rails6_and_up?
-      Rails.application.config.active_record.sqlite3.represent_boolean_as_integer = true
     end
 
     if Devise::Test.rails70?